Restructure API module and introduce Cargo features for webui
Phase 1 - Module restructuring:
- Move src/webui/api.rs to src/api/ (types.rs + handlers.rs)
- Update router.rs imports from crate::webui::api to crate::api
- Remove api re-export from webui/mod.rs

Phase 2 - Cargo features:
- Add [features] section: default=["webui"], webui=["rust-embed","mime_guess"]
- Make rust-embed and mime_guess optional dependencies
- Guard webui module with cfg(feature = "webui") in lib.rs
- Conditionally compile /ui routes and handlers in router.rs
